update 0.0.1
Added the updater. Should only open when the version the backend holds is different from the current app version this means that the app the user has needs to be updated to the most recent version.

Added a way to check the versions in the express backend.

Made the updater have its own window that only opens when it is needed

made other small changes